feednav-data-fetcher/requirements.txt
googlemaps==4.10.0
geopy==2.3.0
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.8.5
